Feature/count habit completion & habit editing (#39)
* fix: balance classes stats

* feat: display user class name and move to inertia page.props instead of prop drilling

* Preloader fade animation on leave

* Footer and Header components

* style: New login page style

* fix: add missing translations

* fix: add translation to footer

* refactor: Add social links to config file

* feat: add hero icons

* Delete unnecessary files

* feat: add missing translations

* fix: change stats with float to int

* chore: update models

* feat: add habit count logic, fix bugs, modify existing logic

* feat-fix: modify & fix logic for handling task counting & passing props from parent comp

* feat: add habit edit modal & logic

* fix: change task tag ids to unsigned int

* feat: add missing translations

* feat: add new routes for habit edit

* fix: change old routes to match new ones

* feat: modify logic of controllers & task service to handle habit edit

+++ database/migrations/2025_03_08_172123_create_task_tag_table.php
@@ -11,8 +11,8 @@ return new class() extends Migration {
     {
         Schema::create("task_tag", function (Blueprint $table): void {
             $table->id();
-            $table->integer("task_id");
-            $table->integer("tag_id");
+            $table->unsignedBigInteger("task_id");
+            $table->unsignedBigInteger("tag_id");
             $table->timestamps();
         });
